Make useValue an alias of useInput
