+++ packages/react-cosmos-renderer/src/client.ts
@@ -2,9 +2,9 @@ export * from './fixture/FixtureCapture/FixtureCapture.js';
 export * from './fixture/FixtureContext.js';
 export * from './fixture/Viewport.js';
 export * from './fixture/useFixtureState.js';
+export * from './fixture/useInput/useInput.js';
+export * from './fixture/useInput/useValue.js';
 export * from './fixture/useSelect/useSelect.js';
-export * from './fixture/useValue/useInput.js';
-export * from './fixture/useValue/useValue.js';
 export * from './fixtureLoaders/ClientFixtureLoader.js';
 export * from './rendererConnect/RendererContext.js';
 export * from './rendererConnect/RendererProvider.js';
