+++ packages/react-cosmos-renderer/src/fixture/useInput/useInput.ts
@@ -1,0 +1,14 @@
+import { useCurrentInputValue } from './useCurrentInputValue.js';
+import { useSetInputValue } from './useSetInputValue.js';
+import { useInputFixtureState } from './useValueFixtureState.js';
+
+export function useInput<T>(
+  inputName: string,
+  defaultValue: T
+): [T, React.Dispatch<React.SetStateAction<T>>] {
+  useInputFixtureState(inputName, defaultValue);
+  const currentValue = useCurrentInputValue(inputName, defaultValue);
+  const setValue = useSetInputValue(inputName, defaultValue);
+
+  return [currentValue, setValue];
+}
