+++ packages/react-cosmos-renderer/src/fixture/useInput/useValueFixtureState.ts
@@ -8,7 +8,7 @@ import {
 } from 'react-cosmos-core';
 import { useFixtureState } from '../useFixtureState.js';
 
-export function useValueFixtureState(inputName: string, defaultValue: unknown) {
+export function useInputFixtureState(inputName: string, defaultValue: unknown) {
   const [, setFixtureState] = useFixtureState<InputsFixtureState>('inputs');
   React.useEffect(() => {
     // The fixture state for this value is (re)created in two situations:
